+++ lib/types/handlebars/index.ts
@@ -1,0 +1,7 @@
+export interface TemplateSpecification {
+    code?: string;
+    map?: string;
+    [key: string | number]: any;
+}
+
+export interface CompileResult extends TemplateSpecification {}